rename UnitConverter.h/.cpp -> Unit.h/.cpp
