+++ source/MaterialXTest/MaterialXCore/Units.cpp
@@ -6,7 +6,7 @@
 #include <MaterialXTest/Catch/catch.hpp>
 
 #include <MaterialXCore/Document.h>
-#include <MaterialXCore/UnitConverter.h>
+#include <MaterialXCore/Unit.h>
 #include <MaterialXFormat/File.h>
 #include <MaterialXFormat/XmlIo.h>
 #include <MaterialXFormat/Util.h>
