+++ source/MaterialXTest/MaterialXGenShader/GenShaderUtil.cpp
@@ -7,7 +7,7 @@
 #include <MaterialXTest/MaterialXGenShader/GenShaderUtil.h>
 
 #include <MaterialXCore/Util.h>
-#include <MaterialXCore/UnitConverter.h>
+#include <MaterialXCore/Unit.h>
 
 #include <MaterialXFormat/File.h>
 #include <MaterialXFormat/Util.h>
